More work on test and start on divide
Now we get an access denied error from the test code, but now we have a
functional block of code that should function well in principles and be
more correct in general. The old arguments were done wrong, that has
been now corrected.
Started also on the divide itself, but it still needs working on.