more revisions to division.py
Iterations are needed widely for this to run without flaw.
The names that come one by one and are removed after use keep the
foldernames from getting ridiculously long.
Without iteration on the create output folders function it made the
outputs look like:
output1
output1output2
output1output2output3
I am generating a function to get output folder's content sizes so they
can be used in the last function, which is dividefilestofolders, which
needs comparison criteria, this makes it easier to compare the
sourcefolder content to a single output folder using rotation over the
output folders.